Replaced protocol and classes to actual files and folders
