+++ Magellan/Classes/Services/MagellanReviewServiceProtocol.swift
